Moving files around to align CovidEquityData
